% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geo_context.R
\name{add_geo_context}
\alias{add_geo_context}
\title{add_geo_context}
\usage{
add_geo_context(x, filepath)
}
\arguments{
\item{x}{text...}

\item{filepath}{text...}
}
\value{

}
\description{
add_geo_context
}
