% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicator_1_7.R
\name{add_indicator_1_7}
\alias{add_indicator_1_7}
\title{add_indicator_1_7}
\usage{
add_indicator_1_7(x, y)
}
\arguments{
\item{x}{tibble containing gwl level head time series}

\item{y}{tibble containing the summary table}
}
\value{

}
\description{
add_indicator_1_7
}
