% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicator_2_3.R
\name{add_indicator_2_3}
\alias{add_indicator_2_3}
\title{add_indicator_2_3}
\usage{
add_indicator_2_3(x, y, selected_reference_period = "Z1")
}
\arguments{
\item{x}{tibble containing gwl level head time series}

\item{y}{tibble containing the summary table}

\item{selected_reference_period}{default}
}
\value{

}
\description{
add_indicator_2_3
}
