% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicator_6_1_3_2.R
\name{add_indicator_6_1_3_2}
\alias{add_indicator_6_1_3_2}
\title{add_indicator_6_1_3_2}
\usage{
add_indicator_6_1_3_2(x, y, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_indicator_6_1_3_2
}
