% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_all.R
\name{add_indicators_all}
\alias{add_indicators_all}
\title{add_indicators_all}
\usage{
add_indicators_all(x, y, ...)
}
\arguments{
\item{x}{A summary table (tibble), that has distinct rows for the columns well_id, climate_model_name and reference_period}

\item{y}{The base table (tibble) containing the time series in long format}

\item{...}{Additional arguments for the functions that calculate the indicators. Currently implemented:
\itemize{
\item selected_reference_period: to definewerwe the reference period
}}
}
\description{
Add all klibiwind7 indicators to a summary table
}
