% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_gwdynindices.R
\name{add_indicators_gwdynindices}
\alias{add_indicators_gwdynindices}
\title{add_indicators_gwdynindices}
\usage{
add_indicators_gwdynindices(x, y)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_indicators_gwdynindices
}
