% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_gwdynindices_shape.R
\name{add_indicators_gwdynindices_shape}
\alias{add_indicators_gwdynindices_shape}
\title{add_indicators_gwdynindices_shape}
\usage{
add_indicators_gwdynindices_shape(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
add_indicators_gwdynindices_shape
}
