% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_gwdynindices_shape_scale.R
\name{add_indicators_gwdynindices_shape_scale}
\alias{add_indicators_gwdynindices_shape_scale}
\title{add_indicators_gwdynindices_shape_scale}
\usage{
add_indicators_gwdynindices_shape_scale(x, y, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_indicators_gwdynindices_shape_scale
}
