% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_gwdynindices_structure.R
\name{add_indicators_gwdynindices_structure}
\alias{add_indicators_gwdynindices_structure}
\title{add_indicators_gwdynindices_structure}
\usage{
add_indicators_gwdynindices_structure(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
add_indicators_gwdynindices_structure
}
