% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indicators_klibiw.R
\name{add_indicators_klibiw}
\alias{add_indicators_klibiw}
\title{add_indicators_klibiw}
\usage{
add_indicators_klibiw(x, y, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
add_indicators_klibiw
}
