% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reference_period_column.R
\name{add_reference_period_column}
\alias{add_reference_period_column}
\title{Add a column with reference period classes}
\usage{
add_reference_period_column(x, reference_period)
}
\arguments{
\item{x}{A tibble with a date column.}

\item{reference_period}{A tibble.}
}
\value{
A tibble x with one additional column containing reference period classes.
}
\description{
Add a column with reference period classes
}
