% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_add.R
\name{circ_add}
\alias{circ_add}
\title{Calculate the circular addition \var{x + y}}
\usage{
circ_add(x, y, int, type = c("minusPiPlusPi", "ZeroPlus2Pi"))
}
\arguments{
\item{circ_minus}{text...}

\item{circular}{text...}
}
\description{
Calculate the circular addition \var{x + y}
}
