% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_minus.R
\name{circ_minus}
\alias{circ_minus}
\title{Calculate the circular subtraction \var{x - y}}
\usage{
circ_minus(x, y, int, type = c("minusPiPlusPi", "ZeroPlus2Pi"))
}
\arguments{
\item{x}{A numeric vector or array.}

\item{y}{A numeric vector or array.}

\item{circular}{}
}
\description{
Calculate the circular subtraction \var{x - y}
}
