% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours_periods.R
\docType{data}
\name{colours_periods}
\alias{colours_periods}
\title{colours_periods}
\format{
\subsection{\code{colours_periods}}{

A vector of length 2 containing hex colour codes.
}
}
\source{
custom definition
}
\usage{
colours_periods
}
\description{
Hex colour codes for the two periods in future
}
\keyword{datasets}
