% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_criterion_incomplete_z1_period.R
\name{filter_criterion_incomplete_z1_period}
\alias{filter_criterion_incomplete_z1_period}
\title{filter_criterion_incomplete_z1_period}
\usage{
filter_criterion_incomplete_z1_period(x)
}
\arguments{
\item{x}{text...}
}
\description{
Drop...
}
