% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_criterion_model_performance.R
\name{filter_criterion_model_performance}
\alias{filter_criterion_model_performance}
\title{filter_criterion_model_performance}
\usage{
filter_criterion_model_performance(x)
}
\arguments{
\item{x}{asda}
}
\description{
Drop...
}
