% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circular_type.R
\name{get_circular_type}
\alias{get_circular_type}
\title{get_circular_type}
\usage{
get_circular_type(x, circ, int, type = c("minusPiPlusPi", "ZeroPlus2Pi"))
}
\arguments{
\item{x}{text...}

\item{circ}{text...}

\item{int}{text...}

\item{type}{text...}
}
\value{

}
\description{
get_circular_type
}
