% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwl_fake_data.R
\docType{data}
\name{gwl_fake_data}
\alias{gwl_fake_data}
\title{gwl_fake_data dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3168 rows and 4 columns.
}
\usage{
gwl_fake_data
}
\description{
This dataframe is used for testing and providing an examle in the readme
}
\author{
Max Nölscher \email{maximilian.noelscher@bgr.de}
}
\keyword{datasets}
