% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_colwell.R
\name{indicator_colwell}
\alias{indicator_colwell}
\title{indicator_colwell}
\usage{
indicator_colwell(
  date,
  gwl,
  fn = "mean",
  s = 11,
  aggregation = "week",
  out = "contingency"
)
}
\arguments{
\item{out}{}
}
\value{

}
\description{
indicator_colwell
}
