% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_highpulseduration.R
\name{indicator_highpulseduration}
\alias{indicator_highpulseduration}
\title{indicator_highpulseduration}
\usage{
indicator_highpulseduration(date, gwl, thr = 0.8)
}
\arguments{
\item{gwl}{}
}
\value{

}
\description{
indicator_highpulseduration
}
