% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_lowpulseduration.R
\name{indicator_lowpulseduration}
\alias{indicator_lowpulseduration}
\title{indicator_lowpulseduration}
\usage{
indicator_lowpulseduration(date, gwl, thr = 0.2)
}
\arguments{
\item{gwl}{}
}
\value{

}
\description{
indicator_lowpulseduration
}
