% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{indicators dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.
}
\usage{
indicators
}
\description{
This dataframe summarises all indicators with their number, label and category
}
\author{
Max Nölscher \email{maximilian.noelscher@bgr.de}
}
\keyword{datasets}
