% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_add_text_to_limits.R
\name{labels_add_text_to_limits}
\alias{labels_add_text_to_limits}
\title{labels_add_text_to_limits}
\usage{
labels_add_text_to_limits(x)
}
\arguments{
\item{x}{character vector containing the labels}
}
\value{

}
\description{
labels_add_text_to_limits
}
