% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_maps_points.R
\name{make_plot_maps_points}
\alias{make_plot_maps_points}
\title{make_plot_maps_points}
\usage{
make_plot_maps_points(plot_data, regions, absolute_value_or_change = "change")
}
\arguments{
\item{plot_data}{text...}

\item{regions}{text...}

\item{absolute_value_or_change}{text...}
}
\value{

}
\description{
make_plot_maps_points
}
