% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_simple_histogram.R
\name{make_plot_simple_histogram}
\alias{make_plot_simple_histogram}
\title{make_plot_simple_histogram}
\usage{
make_plot_simple_histogram(
  plot_data,
  group_var,
  absolute_value_or_change = "change"
)
}
\arguments{
\item{plot_data}{text...}

\item{group_var}{text...}

\item{absolute_value_or_change}{text...}
}
\value{

}
\description{
make_plot_simple_histogram
}
