% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_timeseries_heatmap.R
\name{make_plot_timeseries_heatmap}
\alias{make_plot_timeseries_heatmap}
\title{make_plot_timeseries_heatmap}
\usage{
make_plot_timeseries_heatmap(x, facet_var)
}
\arguments{
\item{x}{text...}

\item{facet_var}{text...}
}
\value{

}
\description{
make_plot_timeseries_heatmap
}
