% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_projections_change_table.R
\name{make_projections_change_table}
\alias{make_projections_change_table}
\title{Title}
\usage{
make_projections_change_table(summary_table)
}
\arguments{
\item{summary_table}{text...}
}
\value{

}
\description{
Title
}
