% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_table.R
\name{make_summary_table}
\alias{make_summary_table}
\title{make_summary_table}
\usage{
make_summary_table(x)
}
\arguments{
\item{x}{The base table (tibble) containing the time series in long format}
}
\description{
Create a summary table
}
