% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_periods.R
\docType{data}
\name{reference_periods}
\alias{reference_periods}
\title{reference_periods dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90 rows and 2 columns.
}
\usage{
reference_periods
}
\description{
This dataframe contains all defined temporal periods with their related years in long format
}
\author{
Max Nölscher \email{maximilian.noelscher@bgr.de}
}
\keyword{datasets}
