% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_periods_fake.R
\docType{data}
\name{reference_periods_fake}
\alias{reference_periods_fake}
\title{reference_periods_fake dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
reference_periods_fake
}
\description{
This dataframe contains all defined temporal periods with their related years in long format for testing and readme
}
\author{
Max Nölscher \email{maximilian.noelscher@bgr.de}
}
\keyword{datasets}
