% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_indicator_3_3.R
\name{unnest_indicator_3_3}
\alias{unnest_indicator_3_3}
\title{unnest_indicator_3_3}
\usage{
unnest_indicator_3_3(x)
}
\arguments{
\item{x}{A summary table (tibble) with the nested indicator 3.3}
}
\description{
Unnest the tibble column of indicator 3.3
}
