% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geo_context.R
\name{add_geo_context}
\alias{add_geo_context}
\title{add_geo_context}
\usage{
add_geo_context(x, filepath, as_sf = FALSE)
}
\arguments{
\item{x}{text...}

\item{filepath}{text...}

\item{as_sf}{text...}
}
\value{

}
\description{
add_geo_context
}
