% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_distribution_vs.R
\name{make_plot_distribution_vs}
\alias{make_plot_distribution_vs}
\title{make_plot_distribution_vs}
\usage{
make_plot_distribution_vs(plot_data, var_rows, var_columns)
}
\arguments{
\item{var_columns}{}
}
\value{

}
\description{
make_plot_distribution_vs
}
