% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_to_yearrange_period_names.R
\name{z_to_yearrange_period_names}
\alias{z_to_yearrange_period_names}
\title{Title}
\usage{
z_to_yearrange_period_names(x_long_format)
}
\arguments{
\item{x_long_format}{}
}
\value{

}
\description{
Title
}
